Reduce fpgm function was moved form analysis.py to subset.py script. BytecodeContainer now is defined on a file of its own.
